# src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "E-Commerce CRM API",
        "description": "High-performance, multi-tenant CRM backend for e-commerce businesses.\n\n## Features\n- Customer management with RFM scoring and lifecycle tracking\n- Order and product catalog management\n- CSV/XLSX import/export with field mapping\n- Shopify integration with webhook sync\n- Role-based access control (RBAC)\n- Background job processing with BullMQ\n\n## Authentication\nAll endpoints require Bearer token authentication. Get a token from `/api/auth/sign-in/email`.\n\n## Rate Limits\n- Auth endpoints: 10 requests/minute\n- API endpoints: 100 requests/minute\n- Bulk operations: 20 requests/minute",
        "version": "2.0.0",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/seifsheikhelarab/E-Commerce-CRM"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main API server"
        }
    ],
...
    "paths": {
...
        "/track/open/{recipientId}": {
            "get": {
                "operationId": "trackEmailOpen",
                "summary": "Track email open event",
                "tags": ["Tracking"],
                "parameters": [
                    {
                        "name": "recipientId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking pixel",
                        "content": { "image/gif": {} }
                    }
                }
            }
        },
        "/track/click/{recipientId}": {
            "get": {
                "operationId": "trackEmailClick",
                "summary": "Track email click event",
                "tags": ["Tracking"],
                "parameters": [
                    {
                        "name": "recipientId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "302": { "description": "Redirect to target URL" }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token from Better Auth sign-in endpoint"
            }
        },
...
    }
}